requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
distlib==0.3.4
dnspython==2.2.1
docopt==0.6.2
filelock==3.7.0
idna==3.3
motor==3.0.0
multidict==6.0.2
parsedatetime==2.6
Pillow==9.1.1
platformdirs==2.5.2
py-cord==2.0.0rc1
pycodestyle==2.8.0
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
toml==0.10.2
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
yarg==0.1.9
yarl==1.7.2
dateparser==1.1.1
